Fixed bug with id numbers and status code on errors. Testing the api.
We always lose a couple of ids.

If I do 100, i get 71 successes and 29 500-errors.
Semaphore was 250, 5, 2, and 1. All have the same results with data returned.

I can only assume either the api has a problem with persistent tcp connections (not true because it serves a website?)
OR there's something else going on. I will try with ids 1-10 and see if there are any missing, and then 1-20. Maybe the api needs a user-agent header or something else going on