Manually closing client.
Every time I make a batch request, there are a few that are not working. I hope to give some respite to the client by managing the disconnection myself, as I notice the last 2 ids are always not going through. I fear this is a temp solution. I will add the user agent string and try a bigger batch for my next test...about 50 ids. This will show if there are more failures scattered through a bigger request