Cleaning up - removing exception stacktrace.
Will now run the full 1400+ batch. And plan the delta load plan later.

Now some questions:
1. Why not a RDBMS? This won't be much data, less than a GB. Can be a sqlite file.
I am doing this because I don't know what analytics I want on top of the data, yet I know I want the data for analytics. We will do an initial load for COMPETITION RESULTS. Next in the model are ATHLETES, then we can get some views going.

I plan to offer not a dataset but a data exploratory platform powered by this data. Therefore I want to try a lakehouse format, with the files served from S3.

2. Doesn't mean it can't be a RDBMS...
True. A managed RDBMS + viewer would be great, or even a site that is built on top of this data....all of that could use relational databases.
Really, the problem is that relational databases are quite mature and have so many protocols built on them for consumption. In the end I want to try a different way, as I have always done things in a relational way.

3. Any other questions?
Well let's see how the queries go.