Running the full initial load.
1422 records.

Took too long - need to identify how to improve the time. However, all future loads will be a delta load so it should be much quicker. Is it worth it to optimise?

Another optimisation technique is skipping all the bad record ids identified. This will make the next run much quicker.