Looks like this solution will work.
Identified a difference in data schema between some records. For an example, between 490 and 491. These two won't be automatically inferred into a Polars dataframe.
The solution is either to give a schema (it is quite nested so difficult and fragile), or to encode it all as a string.
The nested schema is fragile because who knows what other anomalies there are in the data. If you view the data for:
490: New Zealand u15 Championship 2009
491: NZ UNIVERSITY GAMES 2009

Then we see that some events have poules and some don't. Some have Seeding to DE and some don't. All have Final Results as a common factor. Who knows how the schema migrated over the years.

So my thought is to just encode all of this key events as a string. We will transform it in-database as an ELT process once the file lands into the destination.

Common defensive sourcing strategy is there...